+++ src/main/java/com/pentoryall/order/controller/OrderController.java
@@ -1,4 +1,4 @@
-package com.pentoryall.point.controller;
+package com.pentoryall.order.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -7,5 +7,5 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 @RequestMapping
 @RequiredArgsConstructor
-public class TransactionController {
+public class OrderController {
 }
