+++ src/main/java/com/pentoryall/order/dto/ImpOrderDTO.java
@@ -1,4 +1,4 @@
-package com.pentoryall.point.dto;
+package com.pentoryall.order.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,14 +9,11 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 @NoArgsConstructor
-public class TransactionDTO {
+public class ImpOrderDTO {
 
     Long code;
     Long userCode;
-    Long sellerUserCode;
-    Long postCode;
-    Long membershipCode;
-    String type; /* enum? */
+    int amount;
     int point;
     LocalDateTime createDate;
 }
