+++ src/main/java/com/pentoryall/order/mapper/OrderMapper.java
@@ -1,0 +1,7 @@
+package com.pentoryall.order.mapper;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface OrderMapper {
+}
