--- src/main/java/com/pentoryall/point/mapper/OrderMapper.java
@@ -1,10 +1,0 @@
-package com.pentoryall.point.mapper;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface OrderMapper {
-}
